flask/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.4
click==7.1.2
cryptography==3.3.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
google-api-core==2.0.0
google-api-python-client==2.19.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
greenlet==1.1.1
gunicorn==20.0.4
httplib2==0.19.1
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
jwt==1.2.0
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.1
paramiko==2.7.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.23
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==0.16.0
